pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tupianyasuo"
version = "2.0.0"
description = "高质量图片压缩工具 / High-quality image compression tool"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
ui-extras = [
    "ttkbootstrap>=1.10.0",
    "tkinterdnd2>=0.3.0",
]
performance = [
    "pyvips>=2.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "pyinstaller>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long – handled by formatter
    "B008",   # do not perform function calls in default args
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]   # asserts OK in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["ttkbootstrap.*", "tkinterdnd2.*", "pystray.*", "pyvips.*"]
ignore_missing_imports = true
